+++ NEWS.md
@@ -1,5 +1,11 @@
 # ggstatsplot 0.0.11.9000
 
+BUG FIXES
+
+  - Bayes Factor in `ggwithinstats` caption, displayed by default, was
+    incorrect. This has been fixed. This stemmed from a line of code which
+    should have been `paired = TRUE`, but was instead `paired = FALSE`.
+
 MAJOR CHANGES
 
   - The effect size measure for Kruskal-Wallis test has been changed from the